refactor(LottoResult) LottoResult 관련 구조 리팩토링
- 도메인의 의미를 명확히 하기 위헤 클래스명 변경(Budget -> LottoPurchaseBudget, PrizeCalculator -> LottoResult)
- 수익률 계산 로직을 PrizeCalculator에서 LottoPurchaseBudget로 위임
- WinningLotto에서 단일 Ticket의 Prize를 반환하는 calculatePrize 메소드 작성
- 결과 집계를 WinningLotto에서 LottoResult로 위임
- 리팩토링과 클래스 추가로 인한 테스트 추가 및 변경

+++ src/main/java/lotto/domain/dto/LottoResultDto.java
@@ -5,15 +5,15 @@ import lotto.domain.Prize;
 import java.util.Map;
 
 public class LottoResultDto {
-    private final Map<Prize, Integer> lottoResult;
+    private final Map<Prize, Long> lottoResult;
     private final double resultRate;
 
-    public LottoResultDto(Map<Prize, Integer> lottoResult, double resultRate) {
+    public LottoResultDto(Map<Prize, Long> lottoResult, double resultRate) {
         this.lottoResult = lottoResult;
         this.resultRate = resultRate;
     }
 
-    public Map<Prize, Integer> getLottoResult() {
+    public Map<Prize, Long> getLottoResult() {
         return lottoResult;
     }
